Update README.md
# General
##What is this alliance is about?

We aim to create a community to build up a blockchain base system for quick, efficient and traceable help delivery, and supply chain efficiency in times of crisis such as COVID-19

##Why is this project useful?

In time of crisis, it is key to be able to help and support each other in the most efficient way, and to build up trust.

We need to be able to react quickly and make sure that help is delivered as fast as possible in a transparent and traceable way, and that supply chain does not get broken. This project has been created specifically for that purpose. It is an open project to work together toward improving our future and strengthening the link between the people, trust and solidarity.

##How do I get started?
You are welcome to join the discussion in our telegram chat:
https://t.me/mavnAllianceGroup

##Where can I get more help, if I need it?

You might reach out to contact@mavn.ch or post your questions to the Telegram Chat : https://t.me/mavnAllianceGroup